moved send-magic-link OTC tests to send-magic-link.test.js
